Adds the hw module with Computer and MemoryRegion.
A Computer currently holds a vector of MemoryRegion's, which define where and
how a block of memory should be built, and what address offset and size it has.

Over time, many of the hard-coded values in the assembler will make their way
into Computer, or into a sub-struct (Core, maybe).