Fixes ldr instruction error, adds sp register support.
ldr had a copy-paste error from the implementation of str, and the test for ldr
and str didn't do enough to scramble the internal state, so the tests were
passing despite incorrect internal behavior.  The test is better now, and the
defect in ldr (actually the read side of the memory controller) is repaired.

Adds the 'sp' special register to the lexer.

Adds the trivial 'halt' instruction.

Improves the handling of labels in the assembler, and removes an unneded vector
of address, label pairs.